fix pad shape to contain ints not floats

+++ examples/cfq/train.py
@@ -97,7 +97,7 @@ def pad_along_axis(array: jnp.array,
   # The pad function expects a shape of size (n, 2) where n is the number of
   # axes. For each axis, the number of padding positions at the beginning and
   # end of the array is specified.
-  pad_shape = jnp.zeros((array.ndim, 2))
+  pad_shape = jnp.zeros((array.ndim, 2), dtype=jnp.int32)
   pad_shape = jax.ops.index_update(pad_shape,
                                    jax.ops.index[axis, 1],
                                    padding_size)
